+++ build.gradle
@@ -35,35 +35,35 @@ dependencies {
     runtimeOnly 'com.mysql:mysql-connector-j'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
+    // OpenFeign QueryDSL 사용 시
+    implementation "io.github.openfeign.querydsl:querydsl-jpa:6.1"
+    implementation "io.github.openfeign.querydsl:querydsl-core:6.1"
+    annotationProcessor "io.github.openfeign.querydsl:querydsl-apt:6.1:jpa"
+    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+
     // swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.1.0'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-api:2.1.0'
-    // QueryDSL 설정 수정
-    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
-    
-    // APT 설정 변경
-    annotationProcessor(
-        'com.querydsl:querydsl-apt:5.0.0:jakarta',
-        'jakarta.persistence:jakarta.persistence-api',
-        'jakarta.annotation:jakarta.annotation-api'
-    )
-}
-
-// QueryDSL Q클래스 생성 위치 지정
-def querydslDir = "$buildDir/generated/querydsl"
-
-sourceSets {
-    main.java.srcDirs += [ querydslDir ]
-}
-
-tasks.withType(JavaCompile) {
-    options.annotationProcessorGeneratedSourcesDirectory = file(querydslDir)
-}
-
-clean.doLast {
-    file(querydslDir).deleteDir()
 }
 
 tasks.named('test') {
     useJUnitPlatform()
+}
+
+def querydslDir = layout.buildDirectory.dir("generated/querydsl").get().asFile
+
+// 소스 세트에 생성 경로 추가 (구체적인 경로 지정)
+sourceSets {
+    main.java.srcDirs += [ querydslDir ]
+}
+
+// 컴파일 시 생성 경로 지정
+tasks.withType(JavaCompile).configureEach {
+    options.generatedSourceOutputDirectory.set(querydslDir)
+}
+
+// clean 태스크에 생성 폴더 삭제 로직 추가
+clean.doLast {
+    file(querydslDir).deleteDir()
 }
