+++ src/main/java/com/example/springstudy/global/mapping/UserReview.java
@@ -1,7 +1,6 @@
-package com.example.springstudy.mapping;
+package com.example.springstudy.global.mapping;
 
-
-import com.example.springstudy.domain.ask.entity.Ask;
+import com.example.springstudy.domain.review.entity.Review;
 import com.example.springstudy.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
@@ -11,18 +10,18 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Entity
-@Table(name = "user_ask")
-public class UserAsk {
+@Table(name = "user_review")
+public class UserReview {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
     @JoinColumn(name = "user_id")
     private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ask_id")
-    private Ask ask;
+    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
+    @JoinColumn(name = "review_id")
+    private Review review;
 }
